+++ include/nstd/string.h
@@ -196,12 +196,12 @@ NSTDAPI NSTDUInt nstd_string_cap(const NSTDString *string);
 ///
 /// # Returns
 ///
-/// `NSTDErrorCode errc` - Nonzero on error.
+/// `NSTDAllocError errc` - The allocation operation error code.
 ///
 /// # Panics
 ///
 /// Panics if the current length in bytes exceeds `NSTDInt`'s max value.
-NSTDAPI NSTDErrorCode nstd_string_push(NSTDString *string, NSTDUnichar chr);
+NSTDAPI NSTDAllocError nstd_string_push(NSTDString *string, NSTDUnichar chr);
 
 /// Appends a string slice to the end of a string.
 ///